guild_api_runs_cli Show runs as JSON.

IMPORTANT: This command is experimental and subject to change without notice.

IMPORTANT: This command is experimental and subject to change without notice.

Usage

guild_api_runs_cli(
  ...,
  format = FALSE,
  archive = NULL,
  filter = NULL,
  operation = NULL,
  label = NULL,
  unlabeled = FALSE,
  tag = NULL,
  comment = NULL,
  marked = FALSE,
  unmarked = FALSE,
  started = NULL,
  digest = NULL,
  running = FALSE,
  completed = FALSE,
  error = FALSE,
  terminated = FALSE,
  pending = FALSE,
  staged = FALSE,
  api_version = NULL,
  deleted = FALSE,
  include_batch = FALSE
)

guild_api_runs_cli(
  ...,
  format = FALSE,
  archive = NULL,
  filter = NULL,
  operation = NULL,
  label = NULL,
  unlabeled = FALSE,
  tag = NULL,
  comment = NULL,
  marked = FALSE,
  unmarked = FALSE,
  started = NULL,
  digest = NULL,
  running = FALSE,
  completed = FALSE,
  error = FALSE,
  terminated = FALSE,
  pending = FALSE,
  staged = FALSE,
  api_version = NULL,
  deleted = FALSE,
  include_batch = FALSE
)

Arguments

...

passed on to the guild executable. Arguments are automatically quoted with shQuote(), unless they are protected with I(). Pass '--help' or help = TRUE to see all options.

format

(bool) Format the JSON outout.

archive

Show runs archived in PATH.

filter

Filter runs using a filter expression. See Filter by Expression above for details..

operation

Filter runs with operations matching VAL.

label

Filter runs with labels matching VAL. To show unlabeled runs, use unlabeled.

unlabeled

(bool) Filter only runs without labels.

tag

Filter runs with TAG.

comment

Filter runs with comments matching VAL.

marked

(bool) Filter only marked runs.

unmarked

(bool) Filter only unmarked runs.

started

Filter only runs started within RANGE. See above for valid time ranges.

digest

Filter only runs with a matching source code digest.

running

(bool) Filter only runs that are still running.

completed

(bool) Filter only completed runs.

error

(bool) Filter only runs that exited with an error.

terminated

(bool) Filter only runs terminated by the user.

pending

(bool) Filter only pending runs.

staged

(bool) Filter only staged runs.

api_version

API version.

deleted

(bool) Show deleted runs.

include_batch

(bool) Include batch runs.