Copy run files into the current project working directory

Copy run files into the current project working directory

Usage

guild_merge(run = NULL, ...)

Arguments

run

a run selection

...

Arguments passed on to guild_merge_cli

filter

Filter runs using a filter expression. See Filter by Expression above for details.

operation

Filter runs with operations matching VAL.

label

Filter runs with labels matching VAL. To show unlabeled runs, use unlabeled.

unlabeled

(bool) Filter runs without labels.

tag

Filter runs with TAG.

comment

Filter runs with comments matching VAL.

marked

(bool) Filter marked runs.

unmarked

(bool) Filter unmarked runs.

started

Filter runs started within RANGE. See above for valid time ranges.

digest

Filter runs with a matching source code digest.

running

(bool) Filter runs that are still running.

completed

(bool) Filter completed runs.

error

(bool) Filter runs that exited with an error.

terminated

(bool) Filter runs terminated by the user.

pending

(bool) Filter pending runs.

staged

(bool) Filter staged runs.

target_dir

Directory to merge run files to (required if project directory cannot be determined for the run).

sourcecode

(bool) Only copy run source code. Implies use of skip_deps. Cannot be used with skip_sourcecode.

all

(bool) Copy all run files. May be used with skip_sourcecode, skip_deps, and exclude to copy all but the skipped/excluded files.

skip_sourcecode

(bool) Don't copy run source code.

skip_deps

(bool) Don't copy project-local dependencies.

exclude

Exclude a file or pattern (may be used multiple times).

no_summary

(bool) Don't generate a run summary.

summary_name

Name used for the run summary. Use '$run_id' in the name to include the run ID.

preview

(bool) Show what would happen on a merge.

replace

(bool) Allow replacement of existing files. Cannot be used with no_replace

no_replace

(bool) Fail if any target file would be replaced, even if that file is committed to the project VCS. Cannot be used with replace.

Value

NULL, invisibly. This function is called for its side effect.

Examples

if (FALSE) {
guild_merge("--help")
runs_scalars() %>%
  dplyr::slice_max("epoch_acc") %>%
  guild_merge(I("--yes --replace"))
}