guild_merge_cli Copy run files to a project directory.

By default, Guild copies run files into the current directory. To copy files to a different directory, use target_dir DIR.

Usage

guild_merge_cli(
  ...,
  filter = NULL,
  operation = NULL,
  label = NULL,
  unlabeled = NA,
  tag = NULL,
  comment = NULL,
  marked = NA,
  unmarked = NA,
  started = NULL,
  digest = NULL,
  running = NA,
  completed = NA,
  error = NA,
  terminated = NA,
  pending = NA,
  staged = NA,
  target_dir = NULL,
  sourcecode = NA,
  all = NA,
  skip_sourcecode = NA,
  skip_deps = NA,
  exclude = NULL,
  no_summary = NA,
  summary_name = NULL,
  preview = NA,
  replace = NA,
  no_replace = NA
)

Arguments

...

passed on to the guild executable. Arguments are automatically quoted with shQuote(), unless they are protected with I(). Pass '--help' or help = TRUE to see all options.

filter

Filter runs using a filter expression. See Filter by Expression above for details..

operation

Filter runs with operations matching VAL.

label

Filter runs with labels matching VAL. To show unlabeled runs, use unlabeled.

unlabeled

(bool) Filter only runs without labels.

tag

Filter runs with TAG.

comment

Filter runs with comments matching VAL.

marked

(bool) Filter only marked runs.

unmarked

(bool) Filter only unmarked runs.

started

Filter only runs started within RANGE. See above for valid time ranges.

digest

Filter only runs with a matching source code digest.

running

(bool) Filter only runs that are still running.

completed

(bool) Filter only completed runs.

error

(bool) Filter only runs that exited with an error.

terminated

(bool) Filter only runs terminated by the user.

pending

(bool) Filter only pending runs.

staged

(bool) Filter only staged runs.

target_dir

Directory to merge run files to (required if project directory cannot be determined for the run).

sourcecode

(bool) Only copy run source code. Implies use of skip_deps. Cannot be used with skip_sourcecode.

all

(bool) Copy all run files. May be used with skip_sourcecode, skip_deps, and exclude to copy all but the skipped/excluded files.

skip_sourcecode

(bool) Don't copy run source code.

skip_deps

(bool) Don't copy project-local dependencies.

exclude

Exclude a file or pattern (may be used multiple times).

no_summary

(bool) Don't generate a run summary.

summary_name

Name used for the run summary. Use '$run_id' in the name to include the run ID.

preview

(bool) Show what would happen on a merge.

replace

(bool) Allow replacement of existing files. Cannot be used with no_replace

no_replace

(bool) Fail if any target file would be replaced, even if that file is committed to the project VCS. Cannot be used with replace.

Details

Guild checks that the run originated from the current directory before copying files. If the run is associated with a project from a different directory, or is from a package, Guild exits with an error message. In this case, use target_dir to override the check with an explicit path.

The command fails if any file would be replaced, unless a) the replace option is specified or b) the replaced file is committed to the project VCS and unchanged. To prevent replacement even when a file is committed to VCS and unchanged, specify no_replace.