Launch a guild run

Launch a guild run

Usage

guild_run(
  opspec = "train.R",
  flags = NULL,
  ...,
  wait = TRUE,
  label = NULL,
  tags = NULL,
  echo = wait
)

Arguments

opspec

typically path to an R script, but could be an scalar string that guild recognizes as a valid opspec.

flags

- a named list or vector like `c(noise = .3, dropout = .4)`

- a scalar string like `"noise=.3 dropout=.4"`

- a dataframe of flags for a batch of runs

...

Arguments passed on to base::system2

command

the system command to be invoked, as a character string.

args

a character vector of arguments to command.

stdout,stderr

where output to stdout or stderr should be sent. Possible values are "", to the R console (the default), NULL or FALSE (discard output), TRUE (capture the output in a character vector) or a character string naming a file.

stdin

should input be diverted? "" means the default, alternatively a character string naming a file. Ignored if input is supplied.

input

if a character vector is supplied, this is copied one string per line to a temporary file, and the standard input of command is redirected to the file.

env

character vector of name=value strings to set environment variables.

timeout

timeout in seconds, ignored if 0. This is a limit for the elapsed time running command in a separate process. Fractions of seconds are ignored.

minimized,invisible

arguments that are accepted on Windows but ignored on this platform, with a warning.

wait

whether to wait for the run to finish

label, tags

optional strings used to label or tag experiments.

echo

whether output from the run is shown in the current R console. Note, this has no effect on whether expressions are echoed in the guild run stdout.

Value

the return value from `system2()`, invisibly. This function is primarily called for its side effect.