guild_select_cli Select a run and shows its ID.

This command is generally used when specifying a run ID for another Guild command. For example, to restart the latest train run:

Usage

guild_select_cli(
  ...,
  all = FALSE,
  min = NULL,
  max = NULL,
  short_id = FALSE,
  attr = NULL,
  path = FALSE,
  filter = NULL,
  operation = NULL,
  label = NULL,
  unlabeled = FALSE,
  tag = NULL,
  comment = NULL,
  marked = FALSE,
  unmarked = FALSE,
  started = NULL,
  digest = NULL,
  running = FALSE,
  completed = FALSE,
  error = FALSE,
  terminated = FALSE,
  pending = FALSE,
  staged = FALSE
)

Arguments

...

passed on to the guild executable. Arguments are automatically quoted with shQuote(), unless they are protected with I(). Pass '--help' or help = TRUE to see all options.

all

(bool) Select all matching runs, not just the latest.

min

Select the run with the lowest value for the specified COLSPEC.

max

Select the run with the highest value for the specified COLSPEC.

short_id

(bool) Use short ID.

attr

Show specified run attribute rather than run ID.

path

(bool) Show run path.

filter

Filter runs using a filter expression. See Filter by Expression above for details..

operation

Filter runs with operations matching VAL.

label

Filter runs with labels matching VAL. To show unlabeled runs, use unlabeled.

unlabeled

(bool) Filter only runs without labels.

tag

Filter runs with TAG.

comment

Filter runs with comments matching VAL.

marked

(bool) Filter only marked runs.

unmarked

(bool) Filter only unmarked runs.

started

Filter only runs started within RANGE. See above for valid time ranges.

digest

Filter only runs with a matching source code digest.

running

(bool) Filter only runs that are still running.

completed

(bool) Filter only completed runs.

error

(bool) Filter only runs that exited with an error.

terminated

(bool) Filter only runs terminated by the user.

pending

(bool) Filter only pending runs.

staged

(bool) Filter only staged runs.

Details

`guild run --restart $(guild select -o train)`

Specify a Run

You may specify a run using a run ID, a run ID prefix, or a one-based index corresponding to a run returned by the list command.

If RUN isn't specified, the latest matching run is selected.

Selecting Min or Max Scalar

To select the run with the lowest or highest column value, use min or max respectively. For example, to select the run with the lowest loss scalar value, use min loss.

For help with COLSPEC formatting, see COLUMN SPECS in compare help by running guild compare help.

Other run filters are applied before selecting a minimum or maximium scalar value.

Filter by Operation

Runs may be filtered by operation using operation. A run is only included if any part of its full operation name, including the package and model name, matches the value.

Use operation multiple times to include more runs.

Filter by Label

Use label to only include runs with labels containing a specified value. To select runs that do not contain a label, specify a dash '-' for VAL.

Use label multiple times to include more runs.

Filter by Tag

Use tag to only include runs with a specified tag. Tags must match completely and are case sensitive.

Use tag multiple times to include more runs.

Filter by Marked and Unmarked

Use marked to only include marked runs.

Use unmarked to only include unmarked runs. This option may not be used with marked.

Filter by Source Code Digest

To show runs for a specific source code digest, use -g or digest with a complete or partial digest value.